移除依赖
